requirements.txt
alembic==1.6.2
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==5.0.2
six==1.16.0
SQLAlchemy==1.4.14
text-unidecode==1.3
Werkzeug==1.0.1
WTForms==2.3.3
